requirements.txt
Django~=4.2.4
Python-dotenv~=1.0.0
requests~=2.31.0